Note in RVA22U64 that Zihpm was optional in RVA20U64
Section 6.1.2 contains the statement 'The following mandatory extensions
were present in RVA20U64.' followed by a list of extensions, including Zihpm.
This statement is a little confusing as Zihpm is an optional extension in
RVA20U64.

This commit aims to clear up any confusion by adding a note to section 6.2.1
mentioning that Zihpm was optional in RVA20U64.  This is similar to the note added in
rva23-profile.adoc that clarifies the status of 'V' which is optional in RVA22U64
and mandatory in RVA23U64.